- created a separate directory for the modal windows of the project - Moved profile modals and general modals to this folder
