+++ src/components/ModalWindows/SearchModal.vue
