+++ src/components/maps/PlaceDetector.vue
@@ -31,7 +31,7 @@
 <script>
 import { ref } from 'vue'
 
-import ModalWindow from '../../components/ModalWindow.vue'
+import ModalWindow from '../ModalWindows/ModalWindow.vue'
 import PositionMap from '../../components/maps/PositionMap.vue'
 import GreenBtn from '../../components/GreenBtn.vue'
 import Loading from '../../workers/loading-worker/Loading.vue'
