+++ src/components/maps/RegisterModalPositionMap.vue
@@ -28,7 +28,7 @@
 import { ref } from 'vue'
 
 import PositionMap from './PositionMap.vue'
-import ModalWindow from '../ModalWindow.vue'
+import ModalWindow from '../ModalWindows/ModalWindow.vue'
 import GreenBtn from '../../components/GreenBtn.vue'
 import Loading from '../../workers/loading-worker/Loading.vue'
