+++ src/views/application/events/index.vue
@@ -94,7 +94,7 @@ import { v4 as uuid } from 'uuid'
 
 import GreenBtn from '../../../components/GreenBtn.vue'
 import InputComponent from '../../../components/forms/InputComponent.vue'
-import ContextMenu from '../../../components/ContextMenuModal.vue'
+import ContextMenu from '../../../components/ModalWindows/ContextMenuModal.vue'
 import EventCard from '../../../components/event-components/EventCard.vue'
 import SmallLoader from '../../../components/SmallLoader.vue'
 import SearchBlockEvents from '../../../components/SearchBlockEvents.vue'
@@ -473,7 +473,7 @@ export default {
           height: 44px;
           right: 25px;
           z-index: 10;
-          bottom: 130px;
+          bottom: 200px;
 
           @media (max-width: 992px) {
             display: flex;
