+++ src/views/application/events/my-events.vue
@@ -98,7 +98,7 @@ import { v4 as uuid } from 'uuid'
 
 import GreenBtn from '../../../components/GreenBtn.vue'
 import InputComponent from '../../../components/forms/InputComponent.vue'
-import ContextMenu from '../../../components/ContextMenuModal.vue'
+import ContextMenu from '../../../components/ModalWindows/ContextMenuModal.vue'
 import EventCard from '../../../components/event-components/EventCard.vue'
 import SmallLoader from '../../../components/SmallLoader.vue'
 import SearchBlockEvents from '../../../components/SearchBlockEvents.vue'
