+++ src/views/application/index.vue
@@ -57,7 +57,7 @@ import Sidebar from './../../components/Sidebar.vue'
 import MainHeader from './../../components/MainHeader.vue'
 import MobileMenu from '../../components/MobileMenu.vue'
 import Notification from '../../components/Notification.vue'
-import VerifyEmailModal from '../../components/user-cabinet/VerifyEmailModal.vue'
+import VerifyEmailModal from '../../components/ModalWindows/UserCabinetModalWindows/VerifyEmailModal.vue'
 import ModalFeedback from '../../components/ModalFeedback/index.vue'
 import EventCreatedSuccessModal from '../../components/manage-event-components/EventCreatedSuccessModal.vue'
